Move uint/int validator to separate package
This avoids pulling in large dependencies into the public verify
package.
